+++ REPTILE/__init__.py
@@ -2,7 +2,6 @@ from REPTILE.FissionFragmentSpectrum import *
 from REPTILE.EffectiveMass import *
 from REPTILE.ReactionRate import *
 
-from REPTILE.SpectralIndex import *
 from REPTILE.Computables import *
 
 from REPTILE.Calculated import *
